corrigindo para testes

+++ CompraGadosUI/CompraGadosUI/RelatorioCompra.cs
@@ -2,7 +2,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Net;
 using System.Net.Http;
 using System.Windows.Forms;
 
@@ -15,10 +14,12 @@ namespace CompraGadosUI
         int itemsPerPage = 10;
         int numberOfPages = 1;
         CompraGado[] listaCompras;
+        string urlApi;
 
         public RelatorioCompra()
         {
             InitializeComponent();
+            urlApi = Service.GetApiUrl();
         }
 
         private void RelatorioCompra_Load(object sender, EventArgs e)
@@ -29,119 +30,6 @@ namespace CompraGadosUI
         private void btnPesquisar_Click(object sender, EventArgs e)
         {
             Pesquisar();
-        }
-
-        private void Pesquisar()
-        {
-            DateTime dataEntregaDe;
-            DateTime dataEntregaAte;
-            var comboItem = cmbPecuarista.SelectedItem as ComboboxItem;
-            var pecuaristaId = (comboItem == null ? "" : comboItem.Value.ToString());
-
-            if ((IsNullEmptyOrWhiteSpace(txtId.Text)) && (IsNullEmptyOrWhiteSpace(pecuaristaId)) && !DateTime.TryParse(txtDataEntregaDe.Text, out dataEntregaDe) && !DateTime.TryParse(txtDataEntregaAte.Text, out dataEntregaAte))
-            {
-                MessageBox.Show("Pelo menos um dos filtros deve ser preenchido para que seja realizada a pesquisa");
-                return;
-            }
-
-            index = 1;
-            numberOfPages = 0;
-            GetCompras();
-        }
-
-        private async void GetAllPecuarista()
-        {
-            using (var client = new HttpClient())
-            {
-                ServicePointManager.ServerCertificateValidationCallback += (sender, cert, chain, sslPolicyErrors) => true;
-
-                using (var response = await client.GetAsync("http://localhost:5000/api/Pecuarista"))
-                {
-                    if (response.IsSuccessStatusCode)
-                    {
-                        var PecuaristaJsonString = await response.Content.ReadAsStringAsync();
-                        foreach (var item in JsonConvert.DeserializeObject<Pecuarista[]>(PecuaristaJsonString))
-                        {
-                            cmbPecuarista.Items.Add(new ComboboxItem()
-                            {
-                                Text = item.Nome,
-                                Value = item.Id
-                            });
-                        }
-                    }
-                }
-            }
-        }
-
-        private async void GetCompras()
-        {
-            using (var client = new HttpClient())
-            {
-                ServicePointManager.ServerCertificateValidationCallback += (sender, cert, chain, sslPolicyErrors) => true;
-
-                var id = txtId.Text;
-                var pecuaristId = (cmbPecuarista.SelectedItem as ComboboxItem).Value.ToString();
-
-                var url = "http://localhost:5000/api/Compras/Relatorio/" + (IsNullEmptyOrWhiteSpace(id) ? "0" : id) + "/" + (IsNullEmptyOrWhiteSpace(pecuaristId) ? "0" : pecuaristId);
-
-                var sDataEntregaDe = txtDataEntregaDe.Text;
-                DateTime dataEntregaDe;
-                if (DateTime.TryParse(sDataEntregaDe, out dataEntregaDe))
-                {
-                    url += JsonConvert.SerializeObject(dataEntregaDe) + "/";
-                }
-
-                var sDataEntregaAte = txtDataEntregaAte.Text;
-                DateTime dataEntregaAte;
-                if (DateTime.TryParse(sDataEntregaAte, out dataEntregaAte))
-                {
-                    url += JsonConvert.SerializeObject(Convert.ToDateTime(sDataEntregaAte)) + "/";
-                }
-
-                using (var response = await client.GetAsync(url))
-                {
-                    if (response.IsSuccessStatusCode)
-                    {
-                        var JsonString = await response.Content.ReadAsStringAsync();
-                        listaCompras = JsonConvert.DeserializeObject<CompraGado[]>(JsonString);
-
-                        numberOfItems = listaCompras.Count();
-
-                        if (numberOfItems <= itemsPerPage)
-                        {
-                            numberOfPages = 1;
-                        }
-                        else
-                        {
-                            numberOfPages = numberOfItems / itemsPerPage;
-
-                            if (numberOfItems % itemsPerPage > 0)
-                            {
-                                numberOfPages++;
-                            }
-                        }
-
-                        Paginar();
-                    }
-                }
-            }
-        }
-
-        private void Paginar()
-        {
-            gridResultado.DataSource = listaCompras.Skip(itemsPerPage * (index - 1)).Take(itemsPerPage).ToArray();
-
-            gridResultado.Columns["NomePecuarista"].HeaderText = "Pecuarista";
-            gridResultado.Columns["DataEntrega"].HeaderText = "Data Entrega";
-            gridResultado.Columns["ValorTotal"].HeaderText = "Valor Total";
-
-            gridResultado.Columns["PecuaristaId"].Visible = false;
-            gridResultado.Columns["IsImpresso"].Visible = false;
-        }
-
-        private bool IsNullEmptyOrWhiteSpace(string value)
-        {
-            return String.IsNullOrEmpty(value) || String.IsNullOrWhiteSpace(value) || value == null;
         }
 
         private void btnPrevious_Click(object sender, EventArgs e)
@@ -168,24 +56,15 @@ namespace CompraGadosUI
 
         private void btnExcluir_Click(object sender, EventArgs e)
         {
-            var Compra = ObterCompraSelecionada();
-            DeleteCompraGado(Compra.Id);
-        }
+            var Compra = Service.ObterCompraSelecionada<CompraGado>(gridResultado);
 
-        private async void DeleteCompraGado(int id)
-        {
-            using (var client = new HttpClient())
+            if (Compra.IsImpresso)
             {
-                ServicePointManager.ServerCertificateValidationCallback += (sender, cert, chain, sslPolicyErrors) => true;
-
-                using (var response = await client.DeleteAsync("http://localhost:5000/api/Compras/" + id))
-                {
-                    if (response.IsSuccessStatusCode)
-                    {
-                        Pesquisar();
-                    }
-                }
+                MessageBox.Show("A compra está impressa, não pode ser alterada e nem excluída");
+                return;
             }
+
+            DeleteCompraGado(Compra.Id);
         }
 
         private void btnAdicionar_Click(object sender, EventArgs e)
@@ -194,19 +73,9 @@ namespace CompraGadosUI
             ChamarFormCadastro();
         }
 
-        private CompraGado ObterCompraSelecionada()
-        {
-            if (gridResultado.CurrentRow != null)
-            {
-                return (CompraGado)gridResultado.CurrentRow.DataBoundItem;
-            }
-
-            return null;
-        }
-
         private void btnAlterar_Click(object sender, EventArgs e)
         {
-            var Compra = ObterCompraSelecionada();
+            var Compra = Service.ObterCompraSelecionada<CompraGado>(gridResultado);
 
             if (Compra == null)
             {
@@ -218,13 +87,119 @@ namespace CompraGadosUI
             ChamarFormCadastro();
         }
 
+        //Support
         private void ChamarFormCadastro()
         {
             var formCadastro = new CadastroCompra();
             formCadastro.Show();
         }
+
+        private void Paginar()
+        {
+            gridResultado.DataSource = listaCompras.Skip(itemsPerPage * (index - 1)).Take(itemsPerPage).ToArray();
+
+            gridResultado.Columns["NomePecuarista"].HeaderText = "Pecuarista";
+            gridResultado.Columns["DataEntrega"].HeaderText = "Data Entrega";
+            gridResultado.Columns["ValorTotal"].HeaderText = "Valor Total";
+
+            gridResultado.Columns["PecuaristaId"].Visible = false;
+            gridResultado.Columns["IsImpresso"].Visible = false;
+        }
+
+        private void Pesquisar()
+        {
+            DateTime dataEntregaDe;
+            DateTime dataEntregaAte;
+            var comboItem = cmbPecuarista.SelectedItem as ComboboxItem;
+            var pecuaristaId = (comboItem == null ? "" : comboItem.Value.ToString());
+
+            if ((Service.IsNullEmptyOrWhiteSpace(txtId.Text)) && (Service.IsNullEmptyOrWhiteSpace(pecuaristaId)) && !DateTime.TryParse(txtDataEntregaDe.Text, out dataEntregaDe) && !DateTime.TryParse(txtDataEntregaAte.Text, out dataEntregaAte))
+            {
+                MessageBox.Show("Pelo menos um dos filtros deve ser preenchido para que seja realizada a pesquisa");
+                return;
+            }
+
+            index = 1;
+            numberOfPages = 0;
+            GetCompras();
+        }
+
+        // async
+        private async void DeleteCompraGado(int id)
+        {
+            var json = await Service.DeleteApi(Service.GetApiUrl() + "Compras/" + id);
+
+            Pesquisar();
+        }
+
+        private async void GetCompras()
+        {
+            using (var client = new HttpClient())
+            {
+                var id = txtId.Text;
+                var pecuaristId = (cmbPecuarista.SelectedItem as ComboboxItem).Value.ToString();
+
+                var url = "Compras/Relatorio/" + (Service.IsNullEmptyOrWhiteSpace(id) ? "0" : id) + "/" + (Service.IsNullEmptyOrWhiteSpace(pecuaristId) ? "0" : pecuaristId);
+
+                var sDataEntregaDe = txtDataEntregaDe.Text;
+                DateTime dataEntregaDe;
+                if (DateTime.TryParse(sDataEntregaDe, out dataEntregaDe))
+                {
+                    url += JsonConvert.SerializeObject(dataEntregaDe) + "/";
+                }
+
+                var sDataEntregaAte = txtDataEntregaAte.Text;
+                DateTime dataEntregaAte;
+                if (DateTime.TryParse(sDataEntregaAte, out dataEntregaAte))
+                {
+                    url += JsonConvert.SerializeObject(Convert.ToDateTime(sDataEntregaAte)) + "/";
+                }
+
+                var json = await Service.GetApi(Service.GetApiUrl() + url);
+
+                listaCompras = JsonConvert.DeserializeObject<CompraGado[]>(json);
+
+                numberOfItems = listaCompras.Count();
+
+                if (numberOfItems <= itemsPerPage)
+                {
+                    numberOfPages = 1;
+                }
+                else
+                {
+                    numberOfPages = numberOfItems / itemsPerPage;
+
+                    if (numberOfItems % itemsPerPage > 0)
+                    {
+                        numberOfPages++;
+                    }
+                }
+
+                Paginar();
+            }
+        }
+
+        private async void GetAllPecuarista()
+        {
+            var json = await Service.GetApi(Service.GetApiUrl() + "Pecuarista");
+
+            if (Service.IsNullEmptyOrWhiteSpace(json))
+            {
+                return;
+            }
+
+            foreach (var item in JsonConvert.DeserializeObject<Pecuarista[]>(json))
+            {
+                cmbPecuarista.Items.Add(new ComboboxItem()
+                {
+                    Text = item.Nome,
+                    Value = item.Id
+                });
+            }
+        }
     }
 
+    // TODO: chamar Dto
     public class Pecuarista
     {
         public int Id { get; set; }
